Exception de Eventos

+++ event-service/src/main/java/com/grupo1/lucaticket/service/EventServiceImpl.java
@@ -12,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 
@@ -32,7 +33,7 @@ public class EventServiceImpl implements EventService {
 
     @Override
     public List<EventResponse> findAll() {
-        log.info("--- Dentro del método findAll()");
+        log.info("--- Dentro del mÃ©todo findAll()");
         return eventAdapter.of(repository.findAll());
     }
 
@@ -45,6 +46,11 @@ public class EventServiceImpl implements EventService {
     @Override
     public Optional<Event> findById(int id) {
         log.info("***** Antes de encontrar el evento por id");
+        Optional<Event> opt = repository.findById(id);
+        if(opt.isEmpty()) {
+        	log.info("***** No se ha encontrado evento con ese id");
+        	throw new NoSuchElementException("No existe ningun evento con id: "+id);
+        }
         return repository.findById(id);
     }
 
@@ -55,17 +61,32 @@ public class EventServiceImpl implements EventService {
     @Override
     public List<EventResponse> findByNombre(String nombre){
     	log.info("Antes de encontrar el evento por nombre");
+    	List<Event> opt = repository.findByNombre(nombre);
+        if(opt.isEmpty()) {
+        	log.info("***** No se ha encontrado evento con ese nombre");
+        	throw new NoSuchElementException("No existe ningun evento con nombre: "+nombre);
+        }
     	return eventAdapter.of(repository.findByNombre(nombre));
     }
 
 	@Override
 	public List<EventResponse> findByGenero(String genero) {
 		log.info("******* Antes de filtrar por genero");
+		List<Event> opt = repository.findByGenero(genero);
+        if(opt.isEmpty()) {
+        	log.info("***** No se ha encontrado evento con ese genero");
+        	throw new NoSuchElementException("No existe ningun evento con genero: "+genero);
+        }
 		return eventAdapter.of(repository.findByGenero(genero));
 	}
 
 	@Override
 	public void delete(Event event) {
+		Optional<Event> opt = repository.findById(event.getId());
+        if(opt.isEmpty()) {
+        	log.info("***** No se ha encontrado evento con ese genero");
+        	throw new NoSuchElementException("No existe ningun evento con id: "+event.getId());
+        }
 		repository.delete(event);
 		
 	}
